[cocos2dx] Closes #610, closes #627

+++ spine-cocos2dx/src/spine/spine-cocos2dx.cpp
@@ -78,7 +78,7 @@ char* _spUtil_readFile (const char* path, int* length) {
 	Data data = FileUtils::getInstance()->getDataFromFile(
 			FileUtils::getInstance()->fullPathForFilename(path).c_str());
     if (data.isNull()) return 0;
-	*length = data.getSize();
+	*length = static_cast<int>(data.getSize());
 	char* bytes = MALLOC(char, *length);
 	memcpy(bytes, data.getBytes(), *length);
 	return bytes;
